+++ src/main/java/org/riekr/jloga/ui/HoverMonitor.java
@@ -5,20 +5,25 @@ import javax.swing.event.PopupMenuEvent;
 import javax.swing.event.PopupMenuListener;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
+import java.awt.event.WindowEvent;
+import java.awt.event.WindowFocusListener;
 
+import org.riekr.jloga.prefs.Preferences;
 import org.riekr.jloga.react.BoolConsumer;
 
-public class HoverMonitor extends MouseAdapter implements PopupMenuListener {
+public class HoverMonitor extends MouseAdapter implements PopupMenuListener, WindowFocusListener {
 
-	private final Timer   _mouseEnter;
-	private final Timer   _mouseExit;
-	private       boolean _mouseListenerEnabled = true;
+	private final BoolConsumer _consumer;
+	private final Timer        _mouseEnter;
+	private final Timer        _mouseExit;
+	private       boolean      _mouseListenerEnabled = true;
 
 	public HoverMonitor(BoolConsumer consumer) {
 		this(consumer, 200, 300);
 	}
 
 	public HoverMonitor(BoolConsumer consumer, int enterDelay, int exitDelay) {
+		_consumer = consumer;
 		_mouseEnter = new Timer(enterDelay, e -> consumer.accept(true));
 		_mouseEnter.setRepeats(false);
 		_mouseExit = new Timer(exitDelay, e -> consumer.accept(false));
@@ -27,7 +32,7 @@ public class HoverMonitor extends MouseAdapter implements PopupMenuListener {
 
 	@Override
 	public void mouseEntered(MouseEvent e) {
-		if (_mouseListenerEnabled) {
+		if (_mouseListenerEnabled && !Preferences.PRJCLICK.get()) {
 			_mouseExit.stop();
 			_mouseEnter.start();
 		}
@@ -35,10 +40,20 @@ public class HoverMonitor extends MouseAdapter implements PopupMenuListener {
 
 	@Override
 	public void mouseExited(MouseEvent e) {
-		if (_mouseListenerEnabled) {
+		if (_mouseListenerEnabled && !Preferences.PRJCLICK.get()) {
 			_mouseEnter.stop();
 			_mouseExit.start();
 		}
+	}
+
+	@Override
+	public void mouseClicked(MouseEvent e) {
+		if (_mouseListenerEnabled && Preferences.PRJCLICK.get())
+			_consumer.accept(true);
+	}
+
+	public boolean isEnabled() {
+		return _mouseListenerEnabled;
 	}
 
 	public void setEnabled(boolean mouseListenerEnabled) {
@@ -58,5 +73,14 @@ public class HoverMonitor extends MouseAdapter implements PopupMenuListener {
 	@Override
 	public void popupMenuCanceled(PopupMenuEvent e) {
 		_mouseListenerEnabled = true;
+	}
+
+	@Override
+	public void windowGainedFocus(WindowEvent e) {}
+
+	@Override
+	public void windowLostFocus(WindowEvent e) {
+		if (Preferences.PRJCLICK.get())
+			_consumer.accept(false);
 	}
 }